isolate deck dataviews resolver
