+++ libs/deck-layers/src/index.ts
@@ -9,3 +9,4 @@ export * from './layers/fourwings/fourwings.hooks'
 export * from './utils/layers'
 export * from './layer-composer/layer-composer.hooks'
 export * from './layer-composer/types'
+export * from './layer-composer/dataviews-resolver'
